+++ app/src/main/java/ua/rikutou/studio/data/remote/transport/dto/TransportDto.kt
@@ -9,7 +9,7 @@ data class TransportDto(
     val mark: String,
     val manufactureDate: Long,
     val seats: Int,
-    val departmentId: Int,
+    val departmentId: Long,
     val color: String,
     val technicalState: String,
 )
