+++ app/src/main/java/ua/rikutou/studio/data/remote/transport/dto/TransportRequest.kt
@@ -6,7 +6,7 @@ data class TransportRequest(
     val mark: String,
     val manufactureDate: Long,
     val seats: Int,
-    val departmentId: Int,
+    val departmentId: Long,
     val color: String,
     val technicalState: String,
 )
